backend/requirements.txt
Django==5.2.8
djangorestframework==3.15.2
django-cors-headers==4.6.0
Pillow==11.0.0
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
dj-database-url==2.3.0
# psycopg2-binary - Only needed for PostgreSQL in production
# For local development (SQLite), skip this. For production deployment, install separately.
# Windows users: Install with: pip install psycopg2-binary (may need Visual C++ build tools)
whitenoise==6.6.0

